re-adicionando os arquivos certos, e comentei os arquivos que precisa apagar
