+++ src/components/ContentList/index.js
@@ -4,6 +4,11 @@ import ListOfMoviesComponent from './FlatlistComponent';
 
 export default function ContentList({content, endProp, pageNumber}) {
 
+  // switch(numberOfColumns) {
+  //   case numberOfColumns > 500
+
+  // }
+
   return (
     <>
       <FlatList
